+++ R/aggregate.R
@@ -141,9 +141,9 @@ aggregate.sf = function(x, by, FUN, ..., do_union = TRUE, simplify = TRUE,
 #' apply(as.data.frame(nc)[c("BIR74", "BIR79")], 2, sum)
 #' # compare county-wise:
 #' st_agr(bir.grd) = c(BIR74 = "constant")
-#' aw = st_interpolate_aw(bir.grd["BIR74"], st_geometry(nc), extensive = TRUE)
-#' plot(nc$BIR74, aw$BIR74, log = 'xy', xlab = 'county-value', ylab = 'area-w interpolated')
-#' abline(0,1)
+#' \donttest{aw <- st_interpolate_aw(bir.grd["BIR74"], st_geometry(nc), extensive = TRUE)}
+#' \donttest{plot(nc$BIR74, aw$BIR74, log = 'xy', xlab = 'county-value', ylab = 'area-w interpolated')}
+#' \donttest{abline(0,1)}
 #' @export
 st_interpolate_aw = function(x, to, extensive, ...) UseMethod("st_interpolate_aw")
 
@@ -209,6 +209,24 @@ st_interpolate_aw.sf = function(x, to, extensive, ..., keep_NA = FALSE, na.rm = 
 	# clean up:
 	df$...area_t = df$...area_st = df$...area_s = df$Group.1 = NULL
 	st_set_agr(df, "aggregate")
+}
+
+#' @name interpolate_aw
+#' @export
+st_interpolate_aw.stars = function(x, to, extensive, ...) {
+	if (! requireNamespace("stars", quietly = TRUE))
+		stop("package stars required, please install it first")
+	ret = st_interpolate_aw(st_as_sf(x), to, extensive, ...)
+	geom = attr(ret, "sf_column")
+	dx = dim(x)
+	if (length(dx) > 2 && length(x) == 1 && length(ret) > 2) {
+		ret = merge(stars::st_as_stars(ret))
+		nd = names(stars::st_dimensions(x))
+		ret = stars::st_set_dimensions(ret, seq_along(dx), 
+								names = c(geom, paste0(nd[-(1:2)], collapse = ".")))
+		setNames(ret, names(x))
+	} else
+		ret
 }
 
 dasymetric = function(x, to, extensive, keep_NA, include_non_intersected) {