+++ NEWS.md
@@ -1,10 +1,16 @@
 # version 1.1-3
 
+* move `st_interpolate_aw.stars()` from `stars` to `sf`, to get ready for using `GEOS::GEOSGridIntersectionFractions()` in the future
+
 * guess Parquet driver from `.parquet` file extension; #2506
 
 * restore the graphics state after a failed `plot.sf()` key; #2611
 
-* move dependency `classInt` to Suggests:
+* adapt tests to changes in GEOS >= 3.15.0
+
+* new argument `round` to `st_coordinates` to round output coordinates to a given precision level; #2620
+
+* `st_combine()` preserves the precision of input object; #2618
 
 # version 1.1-2
 
@@ -1254,4 +1260,3 @@
 * add and fix `st_agr` API, to set and get attribute-to-geometry relationships
 
 # version 0.2-8, Jan 5, 2017
-